requirements.txt
fastapi==0.115.12
pydantic==2.11.4
pydantic-settings==2.9.1
sqlalchemy==2.0.44
starlette==0.46.0
uvicorn==0.34.2
PyJWT==2.10.1
pwdlib==0.2.1
bcrypt==4.3.0
asyncpg==0.30.0
python-multipart==0.0.20
alembic==1.17.2
httpx==0.28.1
PyYAML==6.0.3
